Fixed spacy python example
